change paddle lite version
